server/requirements/common.txt
APScheduler==2.1.2
diff-match-patch==20121119
Django==1.9.5
django-admin-sortable==2.0.15
django-apptemplates==1.0
django-import-export==0.4.5
https://github.com/samuraisam/django-json-rpc/archive/master.zip#egg=django-json-rpc
django-reversion==1.10.1
Pillow==3.1.1
pytz==2015.7
six==1.9.0
sqlparse==0.1.18
tablib==0.11.1
